+++ doc/release-notes.md
@@ -9,5 +9,6 @@ Wallet Performance Improvements
 
 `zcashd` now reports the following new metrics when `-prometheusport` is set:
 
+- (counter) `zcashd.wallet.batchscanner.outputs.scanned`
 - (gauge) `zcashd.wallet.batchscanner.size.transactions`
 - (gauge) `zcashd.wallet.batchscanner.usage.bytes`